requirements.txt
openai
python-dotenv
langfuse
langsmith
langchain
langchain-core
langchain-openai
langchain-anthropic
requests
tqdm